Flyttet i script
Delt opp js filen.
får opp alle dokumenten i en dropdown meny. Og dokumente som velge blir vist
