+++ versjonsoversikt-poc5/frontend/endring.html
@@ -16,6 +16,7 @@
             <div id="no-selection" class="no-selection">
         </div>
     </div>
-    <script src="js/script.js"></script>
+    <script type="module" scr="js/script.js"></script>
+    <script type="module" src="js/endring.js"></script>
 </body>
 </html>
