+++ versjonsoversikt-poc5/frontend/index.html
@@ -37,6 +37,7 @@
 </form>
 </div>
 
-<script src="js/script.js"></script>
+<script type="module" src="js/script.js"></script>
+<script type="module" src="js/index.js"></script>
 </body>
 </html>
