+++ versjonsoversikt-poc5/frontend/sammenligning.html
@@ -16,17 +16,17 @@
             <div class="content-area pdf-column">
                 <div class="documents-container pdf-block">
                     <h2>Versjon 1</h2>
-                    <select id="pdfSelect1" class="select-sjekkliste pdf-select">
-                        <option value="">Velg PDF...</option>
+                    <select id="document1" class="select-sjekkliste pdf-select">
+                        <option value="">Velg dokument</option>
                     </select>
                     <iframe id="pdfViewer1" class="pdf-viewer" title="PDF 1"></iframe>
                 </div>
 
                 <div class="documents-container pdf-block">
                     <h2>Versjon 2</h2>
-                    <select id="pdfSelect2" class="select-sjekkliste pdf-select">
-                        <option value="">Velg PDF...</option>
-                    </select>
+                    <select id="document2" class="select-sjekkliste pdf-select">
+                        <option value="">Velg dokument</option>
+                    </select> 
                     <iframe id="pdfViewer2" class="pdf-viewer" title="PDF 2"></iframe>
                 </div>
             </div>
@@ -45,6 +45,7 @@
         <div id="statusText" class="status-text"></div>
 
     </div>
-    <script src="js/script.js"></script>
+    <script type="module" src="js/script.js"></script>
+    <script type="module" src="js/sammenligning.js"></script>
 </body>
 </html>
